+++ Sibusten.Philomena.Downloader.Cmd/Commands/CommandExtensions.cs
@@ -24,8 +24,6 @@ namespace Sibusten.Philomena.Downloader.Cmd.Commands
                 new Option<string>(new[] { "--query", "-q" }, "The search query"),
                 new Option<int>(new[] { "--limit", "-l" }, "The maximum number of images to download. Defaults to all images"),
                 new Option<int>(new[] { "--filter", "-f" }, "The ID of the filter to use"),
-                new Option<SortField>(new[] { "--sort-field", "-S" }, "How to sort images"),
-                new Option<SortDirection>(new[] { "--sort-direction", "-D" }, "The direction to sort images"),
                 new Option<string>(new[] { "--image-path", "-I" }, "Where to save images and how to name them"),
                 new Option<string>(new[] { "--json-path", "-J" }, "Where to save json files and how to name them"),
                 new Option<bool?>(new[] { "--skip-images", "-i" }, "Skip saving images"),